Fix that the target objects get through a Player
thats because i mistook "collider" for "collision"...
There also contains a bit of impl. transfer i.e. DamageEffecter.cs to Player.cs
