+++ WirelessAR_Demo/Assets/Scenes/SampleScene.unity
@@ -1195,9 +1195,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _target: {fileID: 1371375972}
-  _camera: {fileID: 1472536871}
-  _isForceForward: 0
-  _speed: 0.005
 --- !u!4 &644640083
 Transform:
   m_ObjectHideFlags: 0
@@ -3009,15 +3006,19 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 1352114948813630, guid: e13ee7e0790c7d243b7aa67fe604acac, type: 3}
       propertyPath: m_TagString
-      value: Player
+      value: Untagged
       objectReference: {fileID: 0}
     - target: {fileID: 1603769914748404, guid: e13ee7e0790c7d243b7aa67fe604acac, type: 3}
       propertyPath: m_Name
-      value: ZED_Rig_Stereo
+      value: Player
       objectReference: {fileID: 0}
     - target: {fileID: 1603769914748404, guid: e13ee7e0790c7d243b7aa67fe604acac, type: 3}
       propertyPath: m_IsActive
       value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1603769914748404, guid: e13ee7e0790c7d243b7aa67fe604acac, type: 3}
+      propertyPath: m_TagString
+      value: Player
       objectReference: {fileID: 0}
     - target: {fileID: 4309805032874704, guid: e13ee7e0790c7d243b7aa67fe604acac, type: 3}
       propertyPath: m_RootOrder
@@ -3110,12 +3111,27 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1472536871}
-  m_Enabled: 0
+  m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 19bd39920667e9344b6e5c81e339ac0b, type: 3}
+  m_Script: {fileID: 11500000, guid: 69ff8b0f04c14c648aafd715d3de9d23, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  img: {fileID: 481780386}
+  _img: {fileID: 481780386}
+  <IsForceForward>k__BackingField: 0
+  <Speed>k__BackingField: 0.01
+--- !u!65 &1472536874
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1472536871}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Size: {x: 0.7, y: 4, z: 0.3}
+  m_Center: {x: 0, y: 0, z: 0.15}
 --- !u!4 &1472536876 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 4309805032874704, guid: e13ee7e0790c7d243b7aa67fe604acac,
@@ -3393,38 +3409,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1690111900}
   m_CullTransparentMesh: 0
---- !u!1 &1694050490 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 1352114948813630, guid: e13ee7e0790c7d243b7aa67fe604acac,
-    type: 3}
-  m_PrefabInstance: {fileID: 1472536868}
-  m_PrefabAsset: {fileID: 0}
---- !u!65 &1694050491
-BoxCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1694050490}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 2
-  m_Size: {x: 0.7, y: 4, z: 0.3}
-  m_Center: {x: 0, y: 0, z: 0.15}
---- !u!114 &1694050492
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1694050490}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 19bd39920667e9344b6e5c81e339ac0b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  img: {fileID: 481780386}
 --- !u!1 &1706361871
 GameObject:
   m_ObjectHideFlags: 0
@@ -3941,7 +3925,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   <Speed>k__BackingField: 30
-  _distance: 1
+  _distance: 3
   _distdata: 0
   <Colors>k__BackingField:
     list:
@@ -3949,7 +3933,7 @@ MonoBehaviour:
       Value: 1
     - Key: {r: 0, g: 0, b: 1, a: 0}
       Value: 1
-  <OrbitDegree>k__BackingField: 16
+  <OrbitDegree>k__BackingField: 0
   <IsRandom>k__BackingField: 0
   <MinOrbitDegree>k__BackingField: -45
   <MaxOrbitDegree>k__BackingField: 45
